pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Tharak-N <tharak.n@innovatily.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
llama-index = "^0.12.5"
fastapi = {extras = ["standard"], version = "^0.115.6"}
langchain = "^0.3.12"
langchain-openai = "^0.2.12"
openai = "^1.57.4"
arize-phoenix = "^7.3.2"
openinference-instrumentation-openai = "^0.1.18"
guidance = "^0.1.16"
scalar-fastapi = "^1.0.3"
botbuilder-core = "^4.16.2"
aiohttp = "^3.11.11"
asyncio = "^3.4.3"
cookiecutter = "^2.6.0"
datetime = "^5.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
